Refactor CLI: fix model receiver naming, remove duplicate import, adjust `helpHeight` initialization, and rename package
